+++ CHANGELOG.md
@@ -10,17 +10,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - `inputMode` to `mwc-textfield` and `mwc-textarea`
 - CSS custom properties for typography
+- Added `autoValidate` property on textfield
 
 ### Changed
 
 - **BREAKING** `--mdc-button-text-transform` has been renamed to `--mdc-typography-button-text-transform`
 - **BREAKING** `--mdc-button-letter-spacing` has been renamed to `--mdc-typography-button-letter-spacing`
 - **BREAKING** `--mdc-tab-text-transform` has been renamed to `--mdc-typography-button-text-transform`
+- **BREAKING:VISUAL** textfield will now only validate on blur instead of input without `autoValidate` prop
 
 ### Fixed
 
 - Setting `scrollTarget` on `mwc-top-app-bar` will update listeners
 - Fixed sass imports of `_index.scss` files
+- Fixed issue with caret jumping to end of input on textfield
 
 ### Changed